finished moving everything to common
