+++ backend/src/api/organization/organization.services.ts
@@ -2,7 +2,8 @@ import { prisma } from "../../utils/prisma";
 import {
   CreateOrganizationBody,
   UpdateOrganizationBody,
-} from "./organization.types";
+} from "common";
+import z from "zod";
 
 export const getOrganization = async (organizationId: string) => {
   const organization = await prisma.organization.findUnique({
@@ -31,7 +32,7 @@ export const getOrganizationFundraisers = async (organizationId: string) => {
 
 // TODO: CHECK FOR POSSIBLE BUG
 export const createOrganization = async (
-  organizationBody: CreateOrganizationBody & { creatorId: string }
+  organizationBody: z.infer<typeof CreateOrganizationBody> & { creatorId: string }
 ) => {
   const newOrganization = await prisma.organization.create({
     data: {
@@ -53,7 +54,7 @@ export const createOrganization = async (
 };
 
 export const updateOrganization = async (
-  organizationBody: UpdateOrganizationBody & { organizationId: string }
+  organizationBody: z.infer<typeof UpdateOrganizationBody> & { organizationId: string }
 ) => {
   const organization = await prisma.organization.update({
     where: { id: organizationBody.organizationId },
