+++ backend/src/api/user/user.handlers.ts
@@ -1,5 +1,6 @@
 import { Request, Response } from "express-serve-static-core";
-import { UpdateUserBody, UserRouteParams } from "./user.types";
+import { UserRouteParams } from "./user.types";
+import { UpdateUserBody } from "common";
 import {
   getUser,
   getUserOrders,
@@ -7,6 +8,7 @@ import {
   updateUser,
 } from "./user.services";
 import { UserSchema, BasicOrderSchema, BasicOrganizationSchema } from "common";
+import z from "zod";
 
 export const getUserHandler = async (
   req: Request<UserRouteParams, any, {}, {}>,
@@ -86,7 +88,7 @@ export const getUserOrganizationsHandler = async (
 };
 
 export const updateUserHandler = async (
-  req: Request<UserRouteParams, any, UpdateUserBody, {}>,
+  req: Request<UserRouteParams, any, z.infer<typeof UpdateUserBody>, {}>,
   res: Response
 ) => {
   if (res.locals.user!.id !== req.params.id) {
