+++ common/schemas/organization.ts
@@ -17,3 +17,24 @@ export const CompleteOrganizationSchema = BasicOrganizationSchema.extend({
 
   admins: z.array(UserSchema),
 });
+
+// CRUD BODIES:
+
+export const CreateOrganizationBody = z.object({
+	name: z.string().min(1).max(255),
+	description: z.string(),
+	logoUrl: z.string().url().optional(),
+	websiteUrl: z.string().url().optional(),
+	instagramUsername: z.string().min(1).max(30).optional(),
+	venmoUsername: z.string().min(5).max(30).optional(),
+});
+
+export const UpdateOrganizationBody = z.object({
+	name: z.string().min(1).max(255).optional(),
+	description: z.string().optional(),
+	logoUrl: z.string().url().optional(),
+	websiteUrl: z.string().url().optional(),
+	instagramUsername: z.string().min(1).max(30).optional(),
+	venmoUsername: z.string().min(5).max(30).optional(),
+	addedAdminsIds: z.array(z.string().uuid()).optional(),
+});
